Fix compile time error in Xcode 13 beta
When compiling in Xcode 13 beta 3 this error occurs:
"'shared' is unavailable in application extensions for iOS: Use view controller based solutions where appropriate instead."

It happens everywhere UIApplication.shared was called.

Solved by marking the classes/extensions/methods as explicitly unavailable in app extensions.
There might be a cleaner way to do this, but seems like this solved the immediate problem.

See: https://developer.apple.com/documentation/xcode-release-notes/xcode-13-beta-release-notes - issue 66928265 for the change in Xcode that introduced the surfacing of this error.